LATTICE-2417 some refactor + finish todo
