+++ src/main/kotlin/com/openlattice/users/Auth0UserListingService.kt
@@ -23,12 +23,7 @@ package com.openlattice.users
 
 import com.auth0.client.mgmt.ManagementAPI
 import com.auth0.json.mgmt.users.User
-import com.dataloom.mappers.ObjectMappers
-import com.geekbeast.auth0.*
 import com.openlattice.users.export.Auth0ApiExtension
-import com.openlattice.users.export.JobStatus
-import com.openlattice.users.export.UserExportJobRequest
-import com.openlattice.users.export.UsersList
 import org.slf4j.LoggerFactory
 import java.time.Instant
 
@@ -45,27 +40,10 @@ class Auth0UserListingService(
 ) : UserListingService {
 
     /**
-     * Calls an export job to download all users from auth0 to a json and parses it to a sequence of users.
-     *
-     * Note: This export is used, because the user API has a 1000 user limitation.
-     * @see <a href="https://auth0.com/docs/users/search/v3/get-users-endpoint#limitations"> Auth0 user endpoint
-     * limitations </a>
+     * Retrieves all users from auth0 as a result of an export job.
      */
     override fun getAllUsers(): Sequence<User> {
-        val exportEntity = auth0ApiExtension.userExport()
-        val job = exportEntity.submitExportJob(
-                UserExportJobRequest(listOf(USER_ID, EMAIL, NICKNAME, APP_METADATA, IDENTITIES))
-        )
-
-        var exportJob = exportEntity.getJob(job.id)
-        while (exportJob.status == JobStatus.pending.name) {
-            exportJob = exportEntity.getJob(job.id)
-        }
-        check(exportJob.status == JobStatus.expired.name) { "Export job ${job.id} expired while trying to retrieve users." }
-
-        val users = exportJob.readUsersList()
-
-        return users.asSequence()
+        return getUsers(auth0ApiExtension).asSequence()
     }
 
     /**
