+++ src/main/kotlin/com/openlattice/users/Auth0Utils.kt
@@ -6,8 +6,11 @@ import com.auth0.exception.Auth0Exception
 import com.auth0.json.mgmt.users.User
 import com.auth0.json.mgmt.users.UsersPage
 import com.auth0.jwt.algorithms.Algorithm
+import com.dataloom.mappers.ObjectMappers
 import com.geekbeast.auth0.*
 import com.openlattice.authentication.Auth0AuthenticationConfiguration
+import com.openlattice.users.export.*
+import org.slf4j.LoggerFactory
 import java.security.InvalidAlgorithmParameterException
 import java.time.Instant
 
@@ -15,12 +18,44 @@ import java.time.Instant
  *
  * @author Matthew Tamayo-Rios &lt;matthew@openlattice.com&gt;
  */
-@Throws(Auth0Exception::class)
-fun getUsers(managementApi: ManagementAPI, page: Int, pageSize: Int): List<User> {
-    require(pageSize <= MAX_PAGE_SIZE) { "Requested page size of $pageSize exceeds max page size of $MAX_PAGE_SIZE " }
 
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+private val logger = LoggerFactory.getLogger("Auth0Utils")
+
+/**
+ * Calls an export job to download all users from auth0 to a json and parses it to a sequence of users.
+ *
+ * Note: This export is used, because the user API has a 1000 user limitation.
+ * @see <a href="https://auth0.com/docs/users/search/v3/get-users-endpoint#limitations"> Auth0 user endpoint
+ * limitations </a>
+ */
+fun getUsers(auth0ApiExtension: Auth0ApiExtension): List<User> {
+    val exportEntity = auth0ApiExtension.userExport()
+    val job = exportEntity.submitExportJob(
+            UserExportJobRequest(listOf(USER_ID, EMAIL, NICKNAME, APP_METADATA, IDENTITIES))
+    )
+
+    var exportJobResult = exportEntity.getJob(job.id)
+    while (exportJobResult.status == JobStatus.pending.name) {
+        exportJobResult = exportEntity.getJob(job.id)
+    }
+
+    return readUsersFromLocation(exportJobResult)
 }
+
+private fun readUsersFromLocation(exportJobResult: UserExportJobResult): UsersList {
+    val downloadUrl = exportJobResult.getDownloadUrl()
+
+    try {
+        return ObjectMappers
+                .getMapper(ObjectMappers.Mapper.valueOf(exportJobResult.format.name))
+                .readValue(downloadUrl.openConnection().getInputStream(), UsersList::class.java)
+    } catch (e: Exception) {
+        logger.error("Couldn't read list of users from download url $downloadUrl.")
+        throw e
+    }
+
+}
+
 
 @Throws(Auth0Exception::class)
 fun getUpdatedUsersPage(
